Added no sensors screen

+++ jetpack_smart_display/app/src/main/java/com/smsoft/smartdisplay/ui/screen/sensors/SensorsScreen.kt
@@ -2,13 +2,22 @@ package com.smsoft.smartdisplay.ui.screen.sensors
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.combinedClickable
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.*
 import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.smsoft.smartdisplay.R
 import com.smsoft.smartdisplay.data.MQTTData
 import com.smsoft.smartdisplay.data.database.entity.emptySensor
 import com.smsoft.smartdisplay.ui.composable.sensors.*
@@ -20,7 +29,7 @@ fun SensorsScreen(
     viewModel: SensorsViewModel = hiltViewModel()
 ) {
     val items by viewModel.getAll.collectAsStateWithLifecycle(
-        initialValue = emptyList()
+        initialValue = null
     )
     var isModifyItems by remember { mutableStateOf(false) }
     var isOpenEditItemDialog by remember { mutableStateOf(false) }
@@ -37,18 +46,28 @@ fun SensorsScreen(
                 }
             ),
         content = { padding ->
-            ItemsList(
-                modifier = Modifier,
-                padding = padding,
-                items = items,
-                itemsData = itemsData,
-                editMode = isModifyItems,
-                onDeleteItem = { item ->
-                    viewModel.deleteItem(item)
+            if (items != null) {
+                if (items!!.isNotEmpty()) {
+                    ItemsList(
+                        modifier = Modifier,
+                        padding = padding,
+                        items = items!!,
+                        itemsData = itemsData,
+                        editMode = isModifyItems,
+                        onDeleteItem = { item ->
+                            viewModel.deleteItem(item)
+                        }
+                    ) { item ->
+                        currentEditItem = item
+                        isOpenEditItemDialog = true
+                    }
+                } else {
+                    NoItems(
+                        modifier = Modifier
+                    ) {
+                        isOpenEditItemDialog = true
+                    }
                 }
-            ) { item ->
-                currentEditItem = item
-                isOpenEditItemDialog = true
             }
             if (isOpenEditItemDialog) {
                 UpdateItem(
@@ -79,4 +98,35 @@ fun SensorsScreen(
             }
         }
     )
+}
+
+@Composable
+fun NoItems(
+    modifier: Modifier,
+    onClick: () -> Unit
+) {
+    Box (
+        modifier = Modifier
+            .fillMaxSize()
+            .pointerInput(Unit){
+                detectTapGestures(
+                    onTap = {
+                        onClick()
+                    }
+                )
+            }
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Text(
+                modifier = Modifier,
+                text = stringResource(R.string.no_sensors_items),
+                color = MaterialTheme.colors.primary
+            )
+        }
+    }
 }
